lesson_29_docker/requirements.txt
psycopg2-binary
pytest
python-dotenv==0.21.1